Breaking the user_mgt.go into multiple files
